requirements.txt
openai
tiktoken
python-dotenv
langchain
chromadb
rich
unstructured
pdf2image
pytesseract